+++ R/dbFixedNames.R
@@ -92,11 +92,7 @@ accClimateVariables <- function(v=NA, domain=NA) {
 #' @return A list where each element is a vector of corresponding country names.
 #' @keywords internal
 getCountryNames <- function(dbname = "gbif4crest_02") {
-    db <- connect_online(dbname = dbname)
-    if(!methods::is(db, 'DBIConnection')) {
-        cat("The connection to the database failed and the process has been stopped. check your internet connection and database IDs.\n")
-        return(NA)
-    }
+    if(!testConnection(dbname)) return(NA)
 
     res <- list()
     req <- "SELECT DISTINCT continent FROM geopolitical_units WHERE continent IS NOT NULL ORDER BY continent"
@@ -149,11 +145,7 @@ accCountryNames <- function(continent=NA) {
 #' @return A list where each element is a vector of corresponding ocean/sea names.
 #' @keywords internal
 getBasinNames <- function(dbname = "gbif4crest_02") {
-    db <- connect_online(dbname = dbname)
-    if(!methods::is(db, 'DBIConnection')) {
-        cat("The connection to the database failed and the process has been stopped. check your internet connection and database IDs.\n")
-        return(NA)
-    }
+    if(!testConnection(dbname)) return(NA)
 
     res <- list()
     req <- "SELECT DISTINCT basin FROM geopolitical_units WHERE basin IS NOT NULL ORDER BY basin"
@@ -212,12 +204,8 @@ accBasinNames <- function(basin=NA) {
 #'         ecoregions) of each realm.
 #' @keywords internal
 getRealmNames <- function(terr = NA, biome = TRUE, ecoregion = TRUE, dbname = "gbif4crest_02") {
-    db <- connect_online(dbname = dbname)
-    if(!methods::is(db, 'DBIConnection')) {
-        cat("The connection to the database failed and the process has been stopped. check your internet connection and database IDs.\n")
-        return(NA)
-    }
-    
+    if(!testConnection(dbname)) return(NA)
+
     res <- list()
     if(is.na(terr)) {
         s <- ''