+++ R/dbGetTaxonomy.R
@@ -28,11 +28,7 @@ getTaxonomy <- function(family = "", genus = "", species = "", taxaType = 1, dep
         stop('No family, genus or species name were provided.\n')
     }
 
-    db <- connect_online(dbname = dbname)
-    if(!methods::is(db, 'DBIConnection')) {
-        cat("The connection to the database failed and the process has been stopped. check your internet connection and database IDs.\n")
-        return(NA)
-    }
+    if(!testConnection(dbname)) return(NA)
 
     family  <- base::trimws(family, 'both')
     genus   <- base::trimws(genus, 'both')