+++ R/plot.explore_calibration_dataset.R
@@ -32,12 +32,7 @@ explore_calibration_dataset <- function( taxaType,
 
     if(base::missing(taxaType)) taxaType
 
-    db <- connect_online(dbname = dbname)
-    if(!methods::is(db, 'DBIConnection')) {
-        cat("[FAILED]\n")
-        cat("The connection to the database failed and the process has been stopped. check your internet connection and database IDs.\n")
-        return(NA)
-    }
+    if(!testConnection(dbname)) return(NA)
 
     coords        <- check_coordinates(xmn, xmx, ymn, ymx)
     xmn           <- coords[1]